+++ src/app/(app)/(personal)/admin/page.tsx
@@ -1,82 +1,26 @@
-import { notFound } from "next/navigation";
-import { getUsersByRole } from "@db/queries";
+import { Suspense } from "react";
 
-import { Role } from "@lib/enums";
-import { getCurrentUser } from "@lib/session";
 import Container from "@components/Container";
 import H1 from "@components/typography/h1";
 import { Separator } from "@components/ui/separator";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@components/ui/tabs";
-import UserCard from "@components/UserCard";
+import { AdminUserTabs, AdminUserTabsSkeleton } from "./AdminUserTabs";
 
 export const metadata = {
   title: "Admin Panel | Wayland Student-Athlete",
   description: "Manage user roles and permissions.",
 };
 
-export default async function AdminDashboard() {
-  const user = await getCurrentUser();
-
-  if (user?.role !== Role.ADMIN) {
-    notFound();
-  }
-
-  const clients = await getUsersByRole(Role.CLIENT);
-  const studentAthletes = await getUsersByRole(Role.STUDENT_ATHLETE);
-  const admins = await getUsersByRole(Role.ADMIN);
-
-  const usersTabs = [
-    {
-      value: "clients",
-      users: clients,
-    },
-    {
-      value: "student-athletes",
-      users: studentAthletes,
-    },
-    {
-      value: "admins",
-      users: admins,
-    },
-  ];
-
+export default function AdminDashboard() {
   return (
     <main className="flex min-h-screen w-full flex-col items-center">
       <Container className="flex size-full max-w-[1600px] flex-col pb-20 pt-32">
-        {user.role === Role.ADMIN ? (
-          <div className="flex flex-col gap-6">
-            <H1>Manage Users</H1>
-            <Separator />
-            <Tabs defaultValue="admins" className="w-full">
-              <TabsList className="mb-4 flex w-full">
-                <TabsTrigger value="clients" className="flex-1">
-                  Clients
-                </TabsTrigger>
-                <TabsTrigger value="student-athletes" className="flex-1">
-                  Student-Athletes
-                </TabsTrigger>
-                <TabsTrigger value="admins" className="flex-1">
-                  Admins
-                </TabsTrigger>
-              </TabsList>
-              {usersTabs.map((tabContent) => (
-                <TabsContent
-                  key={tabContent.value}
-                  value={tabContent.value}
-                  className="grid grid-cols-1 gap-4 md:grid-cols-2 xl:grid-cols-3"
-                >
-                  {tabContent.users.map((user) => (
-                    <UserCard key={user.id} user={user} />
-                  ))}
-                </TabsContent>
-              ))}
-            </Tabs>
-          </div>
-        ) : (
-          <div className="flex flex-col gap-6">
-            <H1>Unauthorized</H1>
-          </div>
-        )}
+        <div className="flex flex-col gap-6">
+          <H1>Manage Users</H1>
+          <Separator />
+          <Suspense fallback={<AdminUserTabsSkeleton />}>
+            <AdminUserTabs />
+          </Suspense>
+        </div>
       </Container>
     </main>
   );
